+++ src/main/kotlin/gwanlang/Expr.kt
@@ -9,4 +9,8 @@ sealed class Expr {
     data class Assign(val name: Token, val value: Expr) : Expr()
     data class Logical(val left: Expr, val op: Token, val right: Expr) : Expr()
     data class Call(val callee: Expr, val paren: Token, val arguments: List<Expr>) : Expr()
+    data class Get(val obj: Expr, val name: Token) : Expr()
+    data class Set(val obj: Expr, val name: Token, val value: Expr) : Expr()
+    data class This(val keyword: Token) : Expr()
+    data class Super(val keyword: Token, val method: Token) : Expr()
 }