+++ src/main/kotlin/gwanlang/GwanFunction.kt
@@ -2,8 +2,15 @@ package gwanlang
 
 class GwanFunction(
     private val declaration: Stmt.Function,
-    private val closure: Environment
+    private val closure: Environment,
+    private val isInitializer: Boolean = false
 ) : GwanCallable {
+
+    fun bind(instance: GwanInstance): GwanFunction {
+        val environment = Environment(closure)
+        environment.define("this", instance)
+        return GwanFunction(declaration, environment, isInitializer)
+    }
 
     override fun arity(): Int = declaration.params.size
 
@@ -15,8 +22,10 @@ class GwanFunction(
         try {
             interpreter.executeBlock(declaration.body, environment)
         } catch (returnValue: Return) {
+            if (isInitializer) return closure.getAt(0, "this")
             return returnValue.value
         }
+        if (isInitializer) return closure.getAt(0, "this")
         return null
     }