+++ src/main/kotlin/gwanlang/Stmt.kt
@@ -9,4 +9,5 @@ sealed class Stmt {
     data class While(val condition: Expr, val body: Stmt) : Stmt()
     data class Function(val name: Token, val params: List<Token>, val body: List<Stmt>) : Stmt()
     data class Return(val keyword: Token, val value: Expr?) : Stmt()
+    data class Class(val name: Token, val superclass: Expr.Variable?, val methods: List<Stmt.Function>) : Stmt()
 }